requirements.txt
crewai
crewai-tools
langchain-community
langchain
langchain-huggingface
langchain-chroma
sentence-transformers
python-dotenv
tavily-python
streamlit
fpdf2
streamlit-mic-recorder
groq
PyPDF2

